Restructured the logic. Added some javadoc.
